feat: add PartialClaudeCodeConfig type for better configuration handling (#115)
# Introduce PartialClaudeCodeConfig type for better type checking

This PR introduces a new `PartialClaudeCodeConfig` type that extends the existing `ClaudeCodeConfig` type. The type signature for the `setup` function and the dev-config have been updated to use this new type, which better represents that configuration options are partial and can be overridden.

This change improves type checking and makes it clearer that the configuration passed to `setup()` doesn't need to include all possible options.